books page hard coding
- card 가 별로 안 이쁨..
- api 로 불러오도록 변경 필요
